(process and signals) Ubuntu programming C
when I am revising for my exams i see this exercise . i tried hard to solve it but i can't . could you help me please? this is the purpose of the activity :

Cars cross a bridge in both directions (direction AB, and direction BA). On the bridge can only pass one line of cars at a time (one direction at a time). Suppose there are two-colored lights on either side of the bridge. We want to simulate the passage of cars on the bridge with the two-tone lights as follows:

1 / Traffic managed by two-colored lights

    The main program runs in a loop.
    The main program creates two threads which play the role of the two bicolor lights (direction AB and direction BA)
    Each passage from one direction to another takes 3 seconds (using SIGALRM).
    At each passage in one direction, the main program sends signals to both wires. Each of them will simulate a two-tone light -The two child processes work as follows:
    A son announces the red light.
    The other son announces the green light.
    After three seconds, a signal is received by each wire and the operation is reversed between the two processes.